FCBHDBP-293 It has added a refactor to avoid the dupicated code. Also, it has added the feature where the plan transformer and playlist transformer should not return the property: translation_item if it is not set.
